test: missing functions and branches
