+++ packages/agents/lighthouse/.nycrc.json
@@ -9,9 +9,13 @@
     "src/adapters/relayer/index.ts",
     "src/tasks/index.ts",
     "src/tasks/run.ts",
+    "src/tasks/prover/handler.ts",
     "src/tasks/prover/adapters/cartographer/index.ts",
     "src/tasks/prover/adapters/index.ts",
+    "src/tasks/prover/bindings/index.ts",
     "src/tasks/prover/operations/index.ts",
+    "src/tasks/prover/operations/consumer.ts",
+    "src/tasks/prover/operations/types.ts",
     "src/tasks/prover/context.ts",
     "src/tasks/prover/index.ts",
     "src/tasks/prover/prover.ts",
