+++ packages/agents/lighthouse/src/tasks/prover/operations/index.ts
@@ -1,2 +1,4 @@
+export * from "./types";
 export { enqueue } from "./publisher";
 export { consume } from "./consumer";
+export { processMessages } from "./process";
